requirements.txt
langchain
streamlit
pypdf

langchain-pinecone
langchain-openai
python-dotenv
langchain-community
beautifulsoup4
streamlit-chat


- e.